I don't think you'll transport more than 65k containers

+++ container_server/headers/cmmp/OutputReadyPacket.hpp
@@ -11,14 +11,14 @@ class OutputReadyPacket : public Packet<OutputReadyPacket> {
 public:
     static const PacketId id;
 
-    OutputReadyPacket(const std::string& license, const std::string& destination, uint32_t capacity)
+    OutputReadyPacket(const std::string& license, const std::string& destination, uint16_t capacity)
         : license(license),
           destination(destination),
           capacity(capacity) {}
 
     const std::string& getLicense() const { return license; }
     const std::string& getDestination() const { return destination; }
-    uint32_t getCapacity() const { return capacity; }
+    uint16_t getCapacity() const { return capacity; }
 
     static OutputReadyPacket decode(std::vector<char>::const_iterator&);
     void encode(std::vector<char>&) const;
@@ -26,7 +26,7 @@ public:
 private:
     std::string license;
     std::string destination;
-    uint32_t capacity;
+    uint16_t capacity;
 
 };
