+++ container_server/src/cmmp/OutputReadyPacket.cpp
@@ -5,7 +5,7 @@ const PacketId OutputReadyPacket::id = PacketId::OutputReady;
 OutputReadyPacket OutputReadyPacket::decode(std::vector<char>::const_iterator& it) {
     std::string license = readString(it);
     std::string destination = readString(it);
-    uint32_t capacity = readPrimitive<uint32_t>(it);
+    uint16_t capacity = readPrimitive<uint16_t>(it);
 
     return OutputReadyPacket(license, destination, capacity);
 }
